app/requirements.txt
click==8.0.3
et-xmlfile==1.1.0
Flask==2.0.2
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
jinjasql==0.1.8
MarkupSafe==2.0.1
numpy==1.22.0
openpyxl==3.0.9
pandas==1.3.5
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
uWSGI==2.0.20
Werkzeug==2.0.2
